support thenable versions of SuperTest
Since v1.1.0, SuperTest ships a version of SuperAgent with "thenable"
support; that is, a then method which provides minimal interop with
Promise libraries but does not provide full then semantics.

See visionmedia/superagent#722 and visionmedia/superagent#726 for
context. Since SuperTest's maintainer is entirely unwilling to support
promises, we'll continue to maintain this library for those who want
`.then()` to return a real promise.